requirements.txt
pymongo
discord.py
cachetools
numpy
dnspython